pieces of deepdeep (low pwr), parking
deepdeep low power uses LPM3 for low power cpu sleeping.  The
problem is this shuts down all clocks other than B CLK which clocks
the RTC hardware and we lose the main system ticker (TA).

We replace lost interrupts using an approximation via the RTC PS h/w
but this is complex and problematic.  So it has been abandoned but
the work is preserved on this branch.